requirements.txt
paramiko
pyaml